// web/graphql-types.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  /** The `ID` scalar type represents a unique identifier, often used to refetch an object or as key for a cache. The ID type appears in a JSON response as a String; however, it is not intended to be human-readable. When expected as an input type, any string (such as `"4"`) or integer (such as `4`) input value will be accepted as an ID. */
  ID: string;
  /** The `String` scalar type represents textual data, represented as UTF-8 character sequences. The String type is most often used by GraphQL to represent free-form human-readable text. */
  String: string;
  /** The `Boolean` scalar type represents `true` or `false`. */
  Boolean: boolean;
  /** The `Int` scalar type represents non-fractional signed whole numeric values. Int can represent values between -(2^31) and 2^31 - 1. */
  Int: number;
  /** The `Float` scalar type represents signed double-precision fractional values as specified by [IEEE 754](https://en.wikipedia.org/wiki/IEEE_floating_point). */
  Float: number;
  /** A date string, such as 2007-12-03, compliant with the ISO 8601 standard for representation of dates and times using the Gregorian calendar. */
  Date: any;
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: any;
};

export type File = Node & {
  sourceInstanceName: Scalars['String'];
  absolutePath: Scalars['String'];
  relativePath: Scalars['String'];
  extension: Scalars['String'];
  size: Scalars['Int'];
  prettySize: Scalars['String'];
  modifiedTime: Scalars['Date'];
  accessTime: Scalars['Date'];
  changeTime: Scalars['Date'];
  birthTime: Scalars['Date'];
  root: Scalars['String'];
  dir: Scalars['String'];
  base: Scalars['String'];
  ext: Scalars['String'];
  name: Scalars['String'];
  relativeDirectory: Scalars['String'];
  dev: Scalars['Int'];
  mode: Scalars['Int'];
  nlink: Scalars['Int'];
  uid: Scalars['Int'];
  gid: Scalars['Int'];
  rdev: Scalars['Int'];
  ino: Scalars['Float'];
  atimeMs: Scalars['Float'];
  mtimeMs: Scalars['Float'];
  ctimeMs: Scalars['Float'];
  atime: Scalars['Date'];
  mtime: Scalars['Date'];
  ctime: Scalars['Date'];
  /** @deprecated Use `birthTime` instead */
  birthtime?: Maybe<Scalars['Date']>;
  /** @deprecated Use `birthTime` instead */
  birthtimeMs?: Maybe<Scalars['Float']>;
  blksize?: Maybe<Scalars['Int']>;
  blocks?: Maybe<Scalars['Int']>;
  id: Scalars['ID'];
  width?: Maybe<Scalars['Int']>;
  height?: Maybe<Scalars['Int']>;
  formats?: Maybe<StrapiNewsletterPromotionHeader_PromoHero_ImageFormats>;
  hash?: Maybe<Scalars['String']>;
  mime?: Maybe<Scalars['String']>;
  url?: Maybe<Scalars['String']>;
  provider?: Maybe<Scalars['String']>;
  provider_metadata?: Maybe<StrapiNewsletterPromotionHeader_PromoHero_ImageProvider_Metadata>;
  created_at?: Maybe<Scalars['Date']>;
  updated_at?: Maybe<Scalars['Date']>;
  alternativeText?: Maybe<Scalars['String']>;
  caption?: Maybe<Scalars['String']>;
  /** Copy file to static directory and return public url to it */
  publicURL?: Maybe<Scalars['String']>;
  /** Returns all children nodes filtered by type ImageSharp */
  childrenImageSharp?: Maybe<Array<Maybe<ImageSharp>>>;
  /** Returns the first child node of type ImageSharp or null if there are no children of given type on this node */
  childImageSharp?: Maybe<ImageSharp>;
  /** Returns all children nodes filtered by type CloudinaryAsset */
  childrenCloudinaryAsset?: Maybe<Array<Maybe<CloudinaryAsset>>>;
  /** Returns the first child node of type CloudinaryAsset or null if there are no children of given type on this node */
  childCloudinaryAsset?: Maybe<CloudinaryAsset>;
  /** Returns all children nodes filtered by type Locale */
  childrenLocale?: Maybe<Array<Maybe<Locale>>>;
  /** Returns the first child node of type Locale or null if there are no children of given type on this node */
  childLocale?: Maybe<Locale>;
  parent?: Maybe<Node>;
  children: Array<Node>;
  internal: Internal;
};
// ...
export type Node = {
  id: Scalars['ID'];
  parent?: Maybe<Node>;
  children: Array<Node>;
  internal: Internal;
};

export type Internal = {
  content?: Maybe<Scalars['String']>;
  contentDigest: Scalars['String'];
  description?: Maybe<Scalars['String']>;
  fieldOwners?: Maybe<Array<Maybe<Scalars['String']>>>;
  ignoreType?: Maybe<Scalars['Boolean']>;
  mediaType?: Maybe<Scalars['String']>;
  owner: Scalars['String'];
  type: Scalars['String'];
};

export type StrapiNewsletterPromotionHeader_PromoHero_ImageFormats = {
  large?: Maybe<StrapiNewsletterPromotionHeader_PromoHero_ImageFormatsLarge>;
  small?: Maybe<StrapiNewsletterPromotionHeader_PromoHero_ImageFormatsSmall>;
  medium?: Maybe<StrapiNewsletterPromotionHeader_PromoHero_ImageFormatsMedium>;
  thumbnail?: Maybe<StrapiNewsletterPromotionHeader_PromoHero_ImageFormatsThumbnail>;
};
// ...
export type StrapiNewsletterPromotionHeader_PromoHero_ImageProvider_Metadata = {
  public_id?: Maybe<Scalars['String']>;
  resource_type?: Maybe<Scalars['String']>;
};
// ...
export type ImageSharp = Node & {
  fixed?: Maybe<ImageSharpFixed>;
  fluid?: Maybe<ImageSharpFluid>;
  gatsbyImageData: Scalars['JSON'];
  original?: Maybe<ImageSharpOriginal>;
  resize?: Maybe<ImageSharpResize>;
  id: Scalars['ID'];
  parent?: Maybe<Node>;
  children: Array<Node>;
  internal: Internal;
};
// ...
export type CloudinaryAsset = Node & {
  fixed: CloudinaryAssetFixed;
  fluid: CloudinaryAssetFluid;
  id: Scalars['ID'];
  parent?: Maybe<Node>;
  children: Array<Node>;
  internal: Internal;
};
// ...
export type StrapiVenue = Node & {
  id: Scalars['ID'];
  name?: Maybe<Scalars['String']>;
  slug_name?: Maybe<Scalars['String']>;
  link?: Maybe<Scalars['String']>;
  lat?: Maybe<Scalars['Float']>;
  lng?: Maybe<Scalars['Float']>;
  website_url?: Maybe<Scalars['String']>;
  cover_image?: Maybe<File>;
  gallery?: Maybe<Array<Maybe<File>>>;
  description?: Maybe<Scalars['String']>;
  isConfirmed?: Maybe<Scalars['Boolean']>;
  address?: Maybe<StrapiVenueAddress>;
  video_url?: Maybe<Scalars['String']>;
  events?: Maybe<Array<Maybe<StrapiFestival>>>;
  videoThumbnail?: Maybe<File>;
  parent?: Maybe<Node>;
  children: Array<Node>;
  internal: Internal;
};

export type StrapiFestival = Node & {
  strapiId?: Maybe<Scalars['Int']>;
  name?: Maybe<Scalars['String']>;
  slug_name?: Maybe<Scalars['String']>;
  description?: Maybe<Scalars['String']>;
  summary?: Maybe<Scalars['String']>;
  credit_cost?: Maybe<Scalars['Int']>;
  cover_image?: Maybe<File>;
  external_img_url?: Maybe<Scalars['String']>;
  gallery?: Maybe<Array<Maybe<File>>>;
  festivalcategories?: Maybe<Array<Maybe<StrapiInterest>>>;
  child_ids?: Maybe<Array<Maybe<StrapiChildIDs>>>;
  capacity?: Maybe<Scalars['Int']>;
  start?: Maybe<Scalars['Date']>;
  end?: Maybe<Scalars['Date']>;
  address?: Maybe<Scalars['String']>;
  location?: Maybe<Scalars['String']>;
  city?: Maybe<Scalars['String']>;
  state?: Maybe<Scalars['String']>;
  country?: Maybe<Scalars['String']>;
  lat?: Maybe<Scalars['Float']>;
  lng?: Maybe<Scalars['Float']>;
  website_url?: Maybe<Scalars['String']>;
  facebook_url?: Maybe<Scalars['String']>;
  contact_email?: Maybe<Scalars['String']>;
  updated_at?: Maybe<Scalars['String']>;
  partner?: Maybe<StrapiPartner>;
  venue?: Maybe<StrapiVenue>;
  lineup_talent?: Maybe<Array<Maybe<StrapiTalent>>>;
  isActive?: Maybe<Scalars['Boolean']>;
  isClaimed?: Maybe<Scalars['Boolean']>;
  isFeatured?: Maybe<Scalars['Boolean']>;
  isConfirmed?: Maybe<Scalars['Boolean']>;
  isParent?: Maybe<Scalars['Boolean']>;
  tickets?: Maybe<Array<Maybe<StrapiProduct>>>;
  video_url?: Maybe<Scalars['String']>;
  videoThumbnail?: Maybe<File>;
  order?: Maybe<Scalars['Int']>;
  amenities?: Maybe<Amenities>;
  id: Scalars['ID'];
  parent?: Maybe<Node>;
  children: Array<Node>;
  internal: Internal;
};

export type Amenities = Node & {
  hasFood?: Maybe<Scalars['Boolean']>;
  hasPrizes?: Maybe<Scalars['Boolean']>;
  hasVipAccess?: Maybe<Scalars['Boolean']>;
  hasRestrooms?: Maybe<Scalars['Boolean']>;
  hasMerchBooth?: Maybe<Scalars['Boolean']>;
  hasFreeParking?: Maybe<Scalars['Boolean']>;
  hasMeetAndGreet?: Maybe<Scalars['Boolean']>;
  hasWheelchairAccess?: Maybe<Scalars['Boolean']>;
  hasPublicTransAccess?: Maybe<Scalars['Boolean']>;
  id: Scalars['ID'];
  parent?: Maybe<Node>;
  children: Array<Node>;
  internal: Internal;
};

export type StrapiFestivalPartner = Node & {
  slug_name?: Maybe<Scalars['String']>;
  id: Scalars['ID'];
  parent?: Maybe<Node>;
  children: Array<Node>;
  internal: Internal;
};

export type StrapiTalent = Node & {
  strapiId?: Maybe<Scalars['Int']>;
  id: Scalars['ID'];
  name?: Maybe<Scalars['String']>;
  slug_name?: Maybe<Scalars['String']>;
  website_url?: Maybe<Scalars['String']>;
  facebook_url?: Maybe<Scalars['String']>;
  instagram_url?: Maybe<Scalars['String']>;
  twitter_url?: Maybe<Scalars['String']>;
  description?: Maybe<Scalars['String']>;
  image?: Maybe<File>;
  gallery?: Maybe<Array<Maybe<File>>>;
  events?: Maybe<Array<Maybe<StrapiFestival>>>;
  parent?: Maybe<Node>;
  children: Array<Node>;
  internal: Internal;
};
// ...
export type StrapiPartner = Node & {
  id: Scalars['ID'];
  name?: Maybe<Scalars['String']>;
  slug_name?: Maybe<Scalars['String']>;
  website_url?: Maybe<Scalars['String']>;
  description?: Maybe<Scalars['String']>;
  logo?: Maybe<File>;
  gallery?: Maybe<Array<Maybe<File>>>;
  events?: Maybe<Array<Maybe<StrapiPartnerEvents>>>;
  isActive?: Maybe<Scalars['Boolean']>;
  organization?: Maybe<Scalars['String']>;
  created_at?: Maybe<Scalars['Date']>;
  updated_at?: Maybe<Scalars['Date']>;
  strapiId?: Maybe<Scalars['Int']>;
  parent?: Maybe<Node>;
  children: Array<Node>;
  internal: Internal;
};
// ...
export type StrapiProduct = Node & {
  interval?: Maybe<Scalars['String']>;
  bonus_tickets?: Maybe<Scalars['Boolean']>;
  display_price?: Maybe<Scalars['String']>;
  credits_awarded?: Maybe<Scalars['Int']>;
  credit_cost?: Maybe<Scalars['Int']>;
  referral_multiplier?: Maybe<Scalars['Float']>;
  stripe_plan_id?: Maybe<Scalars['String']>;
  name?: Maybe<Scalars['String']>;
  description?: Maybe<Scalars['String']>;
  amount?: Maybe<Scalars['Int']>;
  type?: Maybe<Scalars['String']>;
  display_name?: Maybe<Scalars['String']>;
  isActive?: Maybe<Scalars['Boolean']>;
  order?: Maybe<Scalars['Int']>;
  isPrepaidAnnualPlan?: Maybe<Scalars['Boolean']>;
  inboundcampaign?: Maybe<StrapiProductInboundcampaign>;
  outboundcampaigns?: Maybe<StrapiProductOutboundcampaigns>;
  created_at?: Maybe<Scalars['Date']>;
  updated_at?: Maybe<Scalars['Date']>;
  badge?: Maybe<StrapiProductBadge>;
  dynamic_pricing?: Maybe<StrapiProductDynamic_Pricing>;
  strapiId?: Maybe<Scalars['Int']>;
  id: Scalars['ID'];
  giftcode?: Maybe<Scalars['Int']>;
  parent?: Maybe<Node>;
  children: Array<Node>;
  internal: Internal;
};
// ...
export type StrapiInterest = Node & {
  strapiId?: Maybe<Scalars['Int']>;
  name?: Maybe<Scalars['String']>;
  hero_image?: Maybe<File>;
  cover_image?: Maybe<File>;
  isActive?: Maybe<Scalars['Boolean']>;
  created_at?: Maybe<Scalars['Date']>;
  updated_at?: Maybe<Scalars['Date']>;
  parent_id?: Maybe<StrapiInterestParent_Id>;
  isBaseCategory?: Maybe<Scalars['Boolean']>;
  child_ids?: Maybe<Array<Maybe<StrapiInterestChild_Ids>>>;
  id: Scalars['ID'];
  parent?: Maybe<Node>;
  children: Array<Node>;
  internal: Internal;
};
// ...
export type StrapiChildIDs = Node & {
  name?: Maybe<Scalars['String']>;
  slug_name?: Maybe<Scalars['String']>;
  summary?: Maybe<Scalars['String']>;
  cover_image?: Maybe<File>;
  credit_cost?: Maybe<Scalars['String']>;
  start?: Maybe<Scalars['String']>;
  end?: Maybe<Scalars['String']>;
  id: Scalars['ID'];
  parent?: Maybe<Node>;
  children: Array<Node>;
  internal: Internal;
};
// ...
export type Locale = Node & {
  id: Scalars['ID'];
  parent?: Maybe<Node>;
  children: Array<Node>;
  internal: Internal;
  language?: Maybe<Scalars['String']>;
  ns?: Maybe<Scalars['String']>;
  data?: Maybe<Scalars['String']>;
  fileAbsolutePath?: Maybe<Scalars['String']>;
};